rename: GlobalValidator를 validator 패키지로 이동
