+++ src/main/java/lotto/validator/GlobalValidator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.validator;
 
 import java.util.List;
